The kube resolver should set the endpoints when the subsets json structure is not present.
Motivation:

The kube resolver sets the list of endpoints (pods) from the subsets member of the endpoints json object.

When the service does not have any pods, the subsets member is absent and causes an NPE downstream in the resolver assuming the endpoints list is never null.

Changes:

Always set the list of endpoints to consistently populate the list.