Kubernetes service address.
Motivation:

A K8S service sometimes carries more than a single service per endpoint, when a service is resolved by name, the resolver can get multiple ports and cannot determine the port to use.

Changes:

Introduce a Kubernetes service builder that allows to specify the port name or the port number, so the resolver can use it to perform a precise match.